Upload during graceful shutdown
This new feature is disabled by default. However, a slightly
backwards-incompatible change is introduced: in the case that the
KafkaMessageIterator returns false from hasNext (only possible with the legacy
iterator), perform a final upload at the end in this case too (whether or not
the graceful shutdown feature is enabled).